+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Support omitting parentheses in one-line pattern matching. ([@palkan][])
+
 ## 0.13.2 (2021-11-28)
 
 - Parser upgrade.